+++ packages/plugin-packages/mashroom-portal/src/backend/utils/ssr_utils.ts
@@ -2,20 +2,35 @@
 import {createHash} from 'crypto';
 import fetch from 'node-fetch';
 import context from '../context/global_portal_context';
-import {getUser} from './security_utils';
+import {SERVER_SIDE_RENDERED_EMBEDDED_APP_INSTANCE_ID_PREFIX} from '../constants';
+import {getUser, isAppPermitted} from './security_utils';
 import {getResourceAsString} from './resource_utils';
+import {createPortalAppSetup} from './create_portal_app_setup';
 
-import type {Request} from 'express';
+import type {MashroomCDNService} from '@mashroom/mashroom-cdn/type-definitions';
 import type {MashroomLogger} from '@mashroom/mashroom/type-definitions';
+import type {Request} from 'express';
 import type {MashroomMemoryCacheService} from '@mashroom/mashroom-memory-cache/type-definitions';
 import type {
     MashroomPortalAppSetup,
     MashroomPortalService,
     MashroomPortalAppCaching,
     MashroomPortalAppPluginSSRBootstrapFunction,
-    MashroomPortalApp
+    MashroomPortalApp,
+    MashroomPortalAppSSRResult,
+    MashroomPortalAppSSRResultEmbeddedApps,
 } from '../../../type-definitions';
-import type {MashroomPortalIncludeStyleServerSideRenderedAppsResult} from '../../../type-definitions/internal';
+import type {
+    MashroomPortalIncludeStyleServerSideRenderedAppsResult,
+    SSRRenderResult,
+    MashroomPortalPageApps,
+} from '../../../type-definitions/internal';
+
+type RenderEmbeddedPortalAppsResult = {
+    readonly resultHtml: string;
+    readonly embeddedPortalPageApps: MashroomPortalPageApps;
+}
+type RenderEmbeddedPortalAppsFn = (hostHtml: string, portalPageApps: MashroomPortalPageApps) => Promise<RenderEmbeddedPortalAppsResult>;
 
 const CACHE_REGION = 'mashroom-portal-ssr';
 
@@ -33,7 +48,49 @@ const getCacheKey = async (pluginName: string, portalAppSetup: MashroomPortalApp
     return data ? createHash('sha256').update(data).digest('hex') : null;
 };
 
-const renderServerSideWithCache = async (pluginName: string, portalApp: MashroomPortalApp, portalAppSetup: MashroomPortalAppSetup, req: Request, logger: MashroomLogger): Promise<string | null> => {
+const recursivelyRenderEmbeddedApps = async (rootAppHtml: string, embeddedApps: MashroomPortalAppSSRResultEmbeddedApps, renderEmbeddedPortalAppsFn: RenderEmbeddedPortalAppsFn, req: Request, logger: MashroomLogger): Promise<RenderEmbeddedPortalAppsResult> => {
+    const pluginRegistry = context.pluginRegistry;
+    const cdnService: MashroomCDNService | undefined = req.pluginContext.services.cdn?.service;
+    const mashroomSecurityUser = await getUser(req);
+
+    const embeddedPortalPageApps: MashroomPortalPageApps = {};
+    for (const {pluginName, appAreaId, appConfig} of embeddedApps) {
+        logger.debug(`Adding server-side embedded Portal App '${pluginName}' to areaId: ${appAreaId}`);
+
+        const portalApp = pluginRegistry.portalApps.find((app) => app.name === pluginName);
+        if (!portalApp) {
+            logger.error(`Portal App not found: ${pluginName}`);
+            continue;
+        }
+
+        if (!await isAppPermitted(req, pluginName, null, portalApp)) {
+            logger.error(`User '${mashroomSecurityUser ? mashroomSecurityUser.username : 'anonymous'}' is not allowed to access Portal App: ${pluginName}`);
+            continue;
+        }
+
+        const appSetup = await createPortalAppSetup(portalApp, null, appConfig, mashroomSecurityUser, cdnService, pluginRegistry, req);
+
+        embeddedPortalPageApps[appAreaId] = embeddedPortalPageApps[appAreaId] || [];
+        embeddedPortalPageApps[appAreaId].push({
+            pluginName,
+            instanceId: `${SERVER_SIDE_RENDERED_EMBEDDED_APP_INSTANCE_ID_PREFIX}_${appSetup.appId}`,
+            appSetup,
+        });
+    }
+
+    const {resultHtml, embeddedPortalPageApps: embeddedSubPortalApps} = await renderEmbeddedPortalAppsFn(rootAppHtml, embeddedPortalPageApps);
+    Object.keys(embeddedSubPortalApps).forEach((appAreaId) => {
+        embeddedPortalPageApps[appAreaId] = embeddedPortalPageApps[appAreaId] || [];
+        embeddedPortalPageApps[appAreaId].push(...embeddedSubPortalApps[appAreaId]);
+    });
+
+    return {
+        resultHtml,
+        embeddedPortalPageApps,
+    };
+};
+
+const renderServerSideWithCache = async (pluginName: string, portalApp: MashroomPortalApp, portalAppSetup: MashroomPortalAppSetup, renderEmbeddedPortalAppsFn: RenderEmbeddedPortalAppsFn, req: Request, logger: MashroomLogger): Promise<SSRRenderResult | null> => {
     const {cacheEnable, cacheTTLSec} = context.portalPluginConfig.ssrConfig;
     const cacheService: MashroomMemoryCacheService = req.pluginContext.services.memorycache?.service;
     const {ssrBootstrap: ssrBootstrapPath, ssrInitialHtmlUri, cachingConfig} = portalApp;
@@ -44,11 +101,11 @@ const renderServerSideWithCache = async (pluginName: string, portalApp: Mashroom
        const cachedHtml = await cacheService.get(CACHE_REGION, cacheKey);
        if (cachedHtml) {
            logger.debug('SSR cache hit for Portal App', pluginName);
-           return cachedHtml as string;
+           return cachedHtml as SSRRenderResult;
        }
     }
 
-    let html = null;
+    let htmlOrSSRRenderResult: string | MashroomPortalAppSSRResult | null = null;
     if (ssrBootstrapPath) {
         // Local App
         try {
@@ -60,7 +117,7 @@ const renderServerSideWithCache = async (pluginName: string, portalApp: Mashroom
             } else {
                 ssrBootstrap = bootstrap.default;
             }
-            html = await ssrBootstrap(portalAppSetup, req);
+            htmlOrSSRRenderResult = await ssrBootstrap(portalAppSetup, req);
         } catch (e) {
             logger.error(`Loading or executing local SSR bootstrap '${ssrBootstrapPath}' for app '${pluginName}' failed!`, e);
         }
@@ -75,7 +132,12 @@ const renderServerSideWithCache = async (pluginName: string, portalApp: Mashroom
                 headers: {'Content-Type': 'application/json'}
             });
             if (result.ok) {
-                html = await result.text();
+                const contentType = result.headers.get('Content-Type');
+                if (contentType && contentType.indexOf('/json') !== -1) {
+                    htmlOrSSRRenderResult = await result.json();
+                } else {
+                    htmlOrSSRRenderResult = await result.text();
+                }
             } else {
                 logger.error(`Fetching HTML from remote SSR route '${ssrInitialHtmlUri}' for app '${pluginName} failed! HTTP status: ${result.status}`);
             }
@@ -84,15 +146,41 @@ const renderServerSideWithCache = async (pluginName: string, portalApp: Mashroom
         }
     }
 
-    if (html && cacheKey) {
-        // We deliberately don't wait
-        cacheService.set(CACHE_REGION, cacheKey, html, cacheTTLSec);
+    let html;
+    let embeddedApps: MashroomPortalAppSSRResultEmbeddedApps = [];
+    let embeddedPortalPageApps: MashroomPortalPageApps ={};
+
+    if (typeof htmlOrSSRRenderResult === 'string') {
+        html = htmlOrSSRRenderResult;
+    } else if (htmlOrSSRRenderResult && 'html' in htmlOrSSRRenderResult) {
+        html = htmlOrSSRRenderResult.html;
+        embeddedApps = htmlOrSSRRenderResult.embeddedApps ?? [];
     }
 
-    return html;
+    if (!html) {
+        return null;
+    }
+
+    if (embeddedApps.length > 0) {
+        const embeddedAppsRenderResult = await recursivelyRenderEmbeddedApps(html, embeddedApps, renderEmbeddedPortalAppsFn, req, logger);
+        html = embeddedAppsRenderResult.resultHtml;
+        embeddedPortalPageApps = embeddedAppsRenderResult.embeddedPortalPageApps;
+    }
+
+    const ssrRenderResult: SSRRenderResult = {
+        html,
+        embeddedPortalPageApps,
+    };
+
+    if (html && cacheKey) {
+        // We deliberately don't wait
+        cacheService.set(CACHE_REGION, cacheKey, ssrRenderResult, cacheTTLSec);
+    }
+
+    return ssrRenderResult;
 };
 
-export const renderServerSide = async (pluginName: string, portalAppSetup: MashroomPortalAppSetup, req: Request, logger: MashroomLogger): Promise<string | null> => {
+export const renderServerSide = async (pluginName: string, portalAppSetup: MashroomPortalAppSetup, renderEmbeddedPortalAppsFn: RenderEmbeddedPortalAppsFn, req: Request, logger: MashroomLogger, ignoreTimeout = false): Promise<SSRRenderResult | null> => {
     const ssrConfig = context.portalPluginConfig.ssrConfig;
     if (!ssrConfig.ssrEnable) {
         return null;
@@ -104,9 +192,13 @@ export const renderServerSide = async (pluginName: string, portalAppSetup: Mashr
         return null;
     }
 
+    if (ignoreTimeout) {
+        return renderServerSideWithCache(pluginName, portalApp, portalAppSetup, renderEmbeddedPortalAppsFn, req, logger);
+    }
+
     let timeout: any;
     return Promise.race([
-        renderServerSideWithCache(pluginName, portalApp, portalAppSetup, req, logger),
+        renderServerSideWithCache(pluginName, portalApp, portalAppSetup, renderEmbeddedPortalAppsFn, req, logger),
         new Promise<null>((resolve) => timeout = setTimeout(() => resolve(null), ssrConfig.renderTimoutMs)),
     ]).then((result) => {
         clearTimeout(timeout);